refactor: 패키지 추가 및 요금정책 파라미터 수정
