--- src/main/java/nextstep/subway/domain/FareCalculationRule.java
@@ -1,9 +1,0 @@
-package nextstep.subway.domain;
-
-import nextstep.member.domain.Member;
-
-public interface FareCalculationRule {
-    int calculateFare(Path path, Member member, int totalFare);
-
-    void setNextRule(FareCalculationRule nextRule);
-}
