+++ src/main/java/nextstep/subway/domain/fare/FareCalculationRule.java
@@ -1,0 +1,9 @@
+package nextstep.subway.domain.fare;
+
+import nextstep.subway.domain.Path;
+
+public interface FareCalculationRule {
+    int calculateFare(Path path, int age, int totalFare);
+
+    void setNextRule(FareCalculationRule nextRule);
+}
